Создает класс Smart и выносит логику туда
